[Core/Misc]: Restore playercreateinfo_item item deletion functionality.

+++ src/server/game/Globals/ObjectMgr.cpp
@@ -3458,7 +3458,7 @@ void ObjectMgr::LoadPlayerInfo()
                     continue;
                 }
 
-                int32 amount   = fields[3].GetUInt16();
+                int32 amount = fields[3].GetInt8();
 
                 if (!amount)
                 {
